sequences.py
import pandas as pd

def find_longest_sequence(df):
    """
    Finds the longest sequence of rows in a DataFrame that can be visited 
    until the "next" row expires, considering transit, processing, and expiration. 
    The sequence can start at any row in the DataFrame.

    Args:
        df (pd.DataFrame): DataFrame with columns 'ID', 'Init_Transit',
                           'Edge_Transit', and 'Expiration'.

    Returns:
        list: List of IDs of the rows in the longest visitable sequence.
    """

    n_rows = len(df)
    longest_seq = []

    # Function to find the longest sequence starting from a specific row
    def find_sequence_from(start_index):
        curr_seq = []
        time_spent = df.iloc[start_index]['Init_Transit']  # Start with initial transit

        for i in range(start_index, n_rows):
            row = df.iloc[i]

            # Check if we can arrive before expiration
            if time_spent < row['Expiration']:
                time_spent += row['Edge_Transit'] + 120  # Add processing and transit time
                curr_seq.append(row['ID'])
            else:
                break  # Row would expire; end the sequence

        return curr_seq

    # Find longest sequence starting from each row
    for start_index in range(n_rows):
        curr_seq = find_sequence_from(start_index)
        if len(curr_seq) > len(longest_seq):
            longest_seq = curr_seq

    return longest_seq